requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
anyascii==0.3.3
anyio==4.13.0
attrs==26.1.0
audioread==3.1.0
bitsandbytes
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
contourpy==1.3.3
coqpit-config==0.2.5
coqui-tts==0.27.5
coqui-tts-trainer==0.3.3
cuda-bindings==13.2.0
cuda-pathfinder==1.5.3
cuda-toolkit==13.0.2
cycler==0.12.1
decorator==5.2.1
docopt==0.6.2
einops==0.8.2
fastapi
filelock==3.28.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
inflect==7.5.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
ko-speech-tools==0.1.0
lazy-loader==0.5
librosa==0.11.0
llama-cpp-python
llvmlite==0.47.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
monotonic-alignment-search==0.2.1
more-itertools==11.0.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
num2words==0.5.14
numba==0.65.0
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti==13.0.85
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime==13.0.96
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver==12.0.4.66
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse==12.6.3.3
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.20.5
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
nvidia-nvtx-cu12==12.1.105
packaging==26.1
pandas==3.0.2
peft==0.19.1
pillow==12.2.0
platformdirs==4.9.6
pooch==1.9.0
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
pycparser==3.0
Pygments==2.20.0
pyparsing==3.3.2
pysbd==0.3.4
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentencepiece==0.2.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.4.0
torchaudio==2.4.0
tqdm==4.67.3
transformers==5.5.4
triton==3.0.0
typeguard==4.5.1
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn[standard]
Werkzeug==3.1.8
yarl==1.23.0
